requirements.txt
requests>=2.31
python-dotenv>=1.0
fastapi>=0.110
uvicorn[standard]>=0.27
line-bot-sdk>=3.11
